pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codeprep"
version = "0.1.0"
description = "Generate .claude/docs/ knowledge base from git history, static analysis and security audit"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "gitpython>=3.1",
    "rich>=13.0",
]

[project.scripts]
codeprep = "codeprep.cli:cli"

[project.optional-dependencies]
security = ["bandit>=1.7"]
full = ["bandit>=1.7", "tree-sitter>=0.21"]
dev = ["pytest>=7.0", "black", "ruff"]
